Move symlink_missing to the mod_manager directory
